Update: Use new mkdirpStream in dest & symlink
