+++ lib/dest/index.js
@@ -4,8 +4,8 @@ var lead = require('lead');
 var pumpify = require('pumpify');
 var prepare = require('vinyl-prepare');
 
+var fo = require('../file-operations');
 var sourcemap = require('./sourcemap');
-var makeDirs = require('./make-dirs');
 var writeContents = require('./write-contents');
 
 function dest(outFolder, opt) {
@@ -16,7 +16,7 @@ function dest(outFolder, opt) {
   var saveStream = pumpify.obj(
     prepare.dest(outFolder, opt),
     sourcemap(opt),
-    makeDirs(opt),
+    fo.mkdirpStream(opt),
     writeContents(opt)
   );
