+++ lib/file-operations.js
@@ -6,6 +6,8 @@ var fs = require('graceful-fs');
 var path = require('path');
 var assign = require('object-assign');
 var date = require('value-or-function').date;
+var number = require('value-or-function').number;
+var through = require('through2');
 var FlushWriteStream = require('flush-write-stream');
 
 var constants = require('./constants');
@@ -321,6 +323,25 @@ function mkdirp(dirpath, customMode, callback) {
   }
 }
 
+function mkdirpStream(opt) {
+
+  function makeFileDirs(file, enc, callback) {
+    // TODO: Can this be put on file.stat?
+    var dirMode = number(opt.dirMode, file);
+
+    mkdirp(file.dirname, dirMode, onMkdirp);
+
+    function onMkdirp(mkdirpErr) {
+      if (mkdirpErr) {
+        return callback(mkdirpErr);
+      }
+      callback(null, file);
+    }
+  }
+
+  return through.obj(makeFileDirs);
+}
+
 function createWriteStream(path, options, flush) {
   return new WriteStream(path, options, flush);
 }
@@ -427,5 +448,6 @@ module.exports = {
   updateMetadata: updateMetadata,
   writeFile: writeFile,
   mkdirp: mkdirp,
+  mkdirpStream: mkdirpStream,
   createWriteStream: createWriteStream,
 };
