+++ lib/symlink/index.js
@@ -13,7 +13,6 @@ var lead = require('lead');
 
 var fo = require('../file-operations');
 
-var number = valueOrFunction.number;
 var boolean = valueOrFunction.boolean;
 
 var isWindows = (os.platform() === 'win32');
@@ -51,19 +50,7 @@ function symlink(outFolder, opt) {
       srcPath = path.relative(file.base, srcPath);
     }
 
-    // TODO: make DRY with .dest()
-    var dirMode = number(opt.dirMode, file);
-    var writeFolder = path.dirname(file.path);
-
-    fo.mkdirp(writeFolder, dirMode, onMkdirp);
-
-    function onMkdirp(mkdirpErr) {
-      if (mkdirpErr) {
-        return callback(mkdirpErr);
-      }
-
-      fs.symlink(srcPath, file.path, symType, onSymlink);
-    }
+    fs.symlink(srcPath, file.path, symType, onSymlink);
 
     function onSymlink(symlinkErr) {
       if (isErrorFatal(symlinkErr)) {
@@ -75,6 +62,7 @@ function symlink(outFolder, opt) {
 
   var stream = pumpify.obj(
     prepare.dest(outFolder, opt),
+    fo.mkdirpStream(opt),
     through.obj(linkFile)
   );
